+++ Assets/Scripts/Weapons/Weapons/Pole.cs
@@ -7,9 +7,10 @@ public class Pole: SwingWeapon
     [SerializeField]
     private GameObject poleHandle;
     [SerializeField]
-    private float poleDamage = 10f;
+    private int poleDamage = 10;
     [SerializeField]
-    private float poleThrowDamage = 15f;
+    private int poleThrowDamage = 15;
+
 
 	private void Start()
 	{
